Don't add tools_helper dependency to reverie
The tools_helper crate is intended to be used when developing
new reverie tool libraries. Previously, this was being made
available by adding it as a dependency, then making it public via
the reverie library crate.

However, this causes issues in certain cases when building the
binary, because the binary does not implement all the same
functions that would be present in a tool library, leading to
unresolved references.

To solve this, the tools_library is removed as a dependency and
is no longer a default member of the reverie workspace. However,
tool library projects can still include it as a dependency by adding
`tools_helper = { git = "https://github.com/iu-parfunc/reverie" }`
within `Cargo.toml`